#161, #141
Reset animation frames to -1 the first time we encounter an entity.
Then ensure that we do not interpolate the animation in
Cg_AnimateClientEntity. Fixes corpse jitter on respawn, and probably
other animation issues when experiencing packet loss, etc.

+++ src/client/cl_types.h
@@ -47,8 +47,8 @@ typedef struct {
 typedef struct {
 	entity_animation_t animation;
 	uint32_t time;
-	uint16_t frame;
-	uint16_t old_frame;
+	int32_t frame;
+	int32_t old_frame;
 	vec_t lerp;
 	vec_t fraction;
 } cl_entity_animation_t;
